Allow user to reset password (backend impl):
- When user forgets password, user will be able to
reset password, which results in a reset password link
being sent to the user. This link expires in one hour.

- After user clicks link, we verify if the link’s user id
and token id are correct. If correct, user will be able to change
password.

- When user submits new password, we also post the token associated
with the reset process, and reverify if the token is associated
with the user. If it is, we check that the entered password and
confirm password match.

- \TODO (ranger): Impl front end for this process.